`render-promise`
The renderPromise function accepts a component and wraps it in a function that returns a Promise. When this function is called, the component is rendered within an AsyncRoot. The component receives two special props: resolve and reject.

* When the resolve prop is called within the component, the component will be unmounted and the Promise will be resolved with the provided value.

* If the reject prop is called within the component, the Promise will be rejected with the provided error.

This allows for easy rendering of components asynchronously, with the ability to control the flow of execution through Promise resolution and rejection.